requirements.txt
streamlit
scikit-learn
tensorflow
pillow
numpy
joblib
